pyproject.toml
[project]
name = "magic-agents"
version = "0.1.0"
description = "Intent-to-Verification Spec Engine"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "httpx",
    "pyyaml",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
